spliting AdminController, part1 - creating controller for marks
